12.21 JDBC, JPA

+++ h2/src/test/org/h2/test/TestBase.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2004-2022 H2 Group. Multiple-Licensed under the MPL 2.0,
+ * Copyright 2004-2023 H2 Group. Multiple-Licensed under the MPL 2.0,
  * and the EPL 1.0 (https://h2database.com/html/license.html).
  * Initial Developer: H2 Group
  */
@@ -448,7 +448,7 @@ public abstract class TestBase {
         }
         StringUtils.appendTwoDigits(builder, m).append(':');
         StringUtils.appendTwoDigits(builder, s).append('.');
-        StringUtils.appendZeroPadded(builder, 3, millis % 1_000);
+        StringUtils.appendZeroPadded(builder, 3, (int) (millis % 1_000));
         return builder;
     }
